Type Definition atsamd51n::adc0::inputctrl::MUXNEG_R

type MUXNEG_R = R<u8, MUXNEG_A>;

Reader of field MUXNEG

Implementations

impl MUXNEG_R

pub fn variant(&self) -> Variant<u8, MUXNEG_A>

Get enumerated values variant

pub fn is_ain0(&self) -> bool

Checks if the value of the field is AIN0

pub fn is_ain1(&self) -> bool

Checks if the value of the field is AIN1

pub fn is_ain2(&self) -> bool

Checks if the value of the field is AIN2

pub fn is_ain3(&self) -> bool

Checks if the value of the field is AIN3

pub fn is_ain4(&self) -> bool

Checks if the value of the field is AIN4

pub fn is_ain5(&self) -> bool

Checks if the value of the field is AIN5

pub fn is_ain6(&self) -> bool

Checks if the value of the field is AIN6

pub fn is_ain7(&self) -> bool

Checks if the value of the field is AIN7

pub fn is_gnd(&self) -> bool

Checks if the value of the field is GND