Type Definition atsamd51n::dac::dacctrl::OSR_R

type OSR_R = R<u8, OSR_A>;

Reader of field OSR

Implementations

impl OSR_R

pub fn variant(&self) -> Variant<u8, OSR_A>

Get enumerated values variant

pub fn is_osr_1(&self) -> bool

Checks if the value of the field is OSR_1

pub fn is_osr_2(&self) -> bool

Checks if the value of the field is OSR_2

pub fn is_osr_4(&self) -> bool

Checks if the value of the field is OSR_4

pub fn is_osr_8(&self) -> bool

Checks if the value of the field is OSR_8

pub fn is_osr_16(&self) -> bool

Checks if the value of the field is OSR_16

pub fn is_osr_32(&self) -> bool

Checks if the value of the field is OSR_32