Type Definition atsamd51n::dac::dacctrl::REFRESH_R

type REFRESH_R = R<u8, REFRESH_A>;

Reader of field REFRESH

Implementations

impl REFRESH_R

pub fn variant(&self) -> REFRESH_A

Get enumerated values variant

pub fn is_refresh_0(&self) -> bool

Checks if the value of the field is REFRESH_0

pub fn is_refresh_1(&self) -> bool

Checks if the value of the field is REFRESH_1

pub fn is_refresh_2(&self) -> bool

Checks if the value of the field is REFRESH_2

pub fn is_refresh_3(&self) -> bool

Checks if the value of the field is REFRESH_3

pub fn is_refresh_4(&self) -> bool

Checks if the value of the field is REFRESH_4

pub fn is_refresh_5(&self) -> bool

Checks if the value of the field is REFRESH_5

pub fn is_refresh_6(&self) -> bool

Checks if the value of the field is REFRESH_6

pub fn is_refresh_7(&self) -> bool

Checks if the value of the field is REFRESH_7

pub fn is_refresh_8(&self) -> bool

Checks if the value of the field is REFRESH_8

pub fn is_refresh_9(&self) -> bool

Checks if the value of the field is REFRESH_9

pub fn is_refresh_10(&self) -> bool

Checks if the value of the field is REFRESH_10

pub fn is_refresh_11(&self) -> bool

Checks if the value of the field is REFRESH_11

pub fn is_refresh_12(&self) -> bool

Checks if the value of the field is REFRESH_12

pub fn is_refresh_13(&self) -> bool

Checks if the value of the field is REFRESH_13

pub fn is_refresh_14(&self) -> bool

Checks if the value of the field is REFRESH_14

pub fn is_refresh_15(&self) -> bool

Checks if the value of the field is REFRESH_15