Type Definition atsamd51n::dmac::channel::chctrla::THRESHOLD_R

type THRESHOLD_R = R<u8, THRESHOLD_A>;

Reader of field THRESHOLD

Implementations

impl THRESHOLD_R

pub fn variant(&self) -> THRESHOLD_A

Get enumerated values variant

pub fn is_1beat(&self) -> bool

Checks if the value of the field is _1BEAT

pub fn is_2beats(&self) -> bool

Checks if the value of the field is _2BEATS

pub fn is_4beats(&self) -> bool

Checks if the value of the field is _4BEATS

pub fn is_8beats(&self) -> bool

Checks if the value of the field is _8BEATS