Type Definition atsamd51n::dsu::cfg::DCCDMALEVEL_R

type DCCDMALEVEL_R = R<u8, DCCDMALEVEL_A>;

Reader of field DCCDMALEVEL

Implementations

impl DCCDMALEVEL_R

pub fn variant(&self) -> Variant<u8, DCCDMALEVEL_A>

Get enumerated values variant

pub fn is_empty(&self) -> bool

Checks if the value of the field is EMPTY

pub fn is_full(&self) -> bool

Checks if the value of the field is FULL