Type Definition atsamd51n::evsys::channel::channel::PATH_R

type PATH_R = R<u8, PATH_A>;

Reader of field PATH

Implementations

impl PATH_R

pub fn variant(&self) -> Variant<u8, PATH_A>

Get enumerated values variant

pub fn is_synchronous(&self) -> bool

Checks if the value of the field is SYNCHRONOUS

pub fn is_resynchronized(&self) -> bool

Checks if the value of the field is RESYNCHRONIZED

pub fn is_asynchronous(&self) -> bool

Checks if the value of the field is ASYNCHRONOUS