Type Definition atsamd51n::gclk::genctrl::SRC_R

type SRC_R = R<u8, SRC_A>;

Reader of field SRC

Implementations

impl SRC_R

pub fn variant(&self) -> Variant<u8, SRC_A>

Get enumerated values variant

pub fn is_xosc0(&self) -> bool

Checks if the value of the field is XOSC0

pub fn is_xosc1(&self) -> bool

Checks if the value of the field is XOSC1

pub fn is_gclkin(&self) -> bool

Checks if the value of the field is GCLKIN

pub fn is_gclkgen1(&self) -> bool

Checks if the value of the field is GCLKGEN1

pub fn is_osculp32k(&self) -> bool

Checks if the value of the field is OSCULP32K

pub fn is_xosc32k(&self) -> bool

Checks if the value of the field is XOSC32K

pub fn is_dfll(&self) -> bool

Checks if the value of the field is DFLL

pub fn is_dpll0(&self) -> bool

Checks if the value of the field is DPLL0

pub fn is_dpll1(&self) -> bool

Checks if the value of the field is DPLL1