Type Definition atsamd51n::i2s::rxctrl::EXTEND_R

type EXTEND_R = R<u8, EXTEND_A>;

Reader of field EXTEND

Implementations

impl EXTEND_R

pub fn variant(&self) -> EXTEND_A

Get enumerated values variant

pub fn is_zero(&self) -> bool

Checks if the value of the field is ZERO

pub fn is_one(&self) -> bool

Checks if the value of the field is ONE

pub fn is_msbit(&self) -> bool

Checks if the value of the field is MSBIT

pub fn is_lsbit(&self) -> bool

Checks if the value of the field is LSBIT