Type Definition atsamd51n::nvmctrl::ctrla::PRM_R

type PRM_R = R<u8, PRM_A>;

Reader of field PRM

Implementations

impl PRM_R

pub fn variant(&self) -> Variant<u8, PRM_A>

Get enumerated values variant

pub fn is_semiauto(&self) -> bool

Checks if the value of the field is SEMIAUTO

pub fn is_fullauto(&self) -> bool

Checks if the value of the field is FULLAUTO

pub fn is_manual(&self) -> bool

Checks if the value of the field is MANUAL