Struct atsamd51n::nvmctrl::ctrlb::CMD_W

pub struct CMD_W<'a> { /* fields omitted */ }

Write proxy for field CMD

Implementations

impl<'a> CMD_W<'a>

pub fn variant(self, variant: CMD_AW) -> &'a mut W

Writes variant to the field

pub fn ep(self) -> &'a mut W

Erase Page - Only supported in the USER and AUX pages.

pub fn eb(self) -> &'a mut W

Erase Block - Erases the block addressed by the ADDR register, not supported in the user page

pub fn wp(self) -> &'a mut W

Write Page - Writes the contents of the page buffer to the page addressed by the ADDR register, not supported in the user page

pub fn wqw(self) -> &'a mut W

Write Quad Word - Writes a 128-bit word at the location addressed by the ADDR register.

pub fn swrst(self) -> &'a mut W

Software Reset - Power-Cycle the NVM memory and replay the device automatic calibration procedure and resets the module configuration registers

pub fn lr(self) -> &'a mut W

Lock Region - Locks the region containing the address location in the ADDR register.

pub fn ur(self) -> &'a mut W

Unlock Region - Unlocks the region containing the address location in the ADDR register.

pub fn sprm(self) -> &'a mut W

Sets the power reduction mode.

pub fn cprm(self) -> &'a mut W

Clears the power reduction mode.

pub fn pbc(self) -> &'a mut W

Page Buffer Clear - Clears the page buffer.

pub fn ssb(self) -> &'a mut W

Set Security Bit

pub fn bkswrst(self) -> &'a mut W

Bank swap and system reset, if SMEE is used also reallocate SMEE data into the opposite BANK

pub fn celck(self) -> &'a mut W

Chip Erase Lock - DSU.CE command is not available

pub fn ceulck(self) -> &'a mut W

Chip Erase Unlock - DSU.CE command is available

pub fn sbpdis(self) -> &'a mut W

Sets STATUS.BPDIS, Boot loader protection is discarded until CBPDIS is issued or next start-up sequence

pub fn cbpdis(self) -> &'a mut W

Clears STATUS.BPDIS, Boot loader protection is not discarded

pub fn asees0(self) -> &'a mut W

Activate SmartEEPROM Sector 0, deactivate Sector 1

pub fn asees1(self) -> &'a mut W

Activate SmartEEPROM Sector 1, deactivate Sector 0

pub fn seeraloc(self) -> &'a mut W

Starts SmartEEPROM sector reallocation algorithm

pub fn seeflush(self) -> &'a mut W

Flush SMEE data when in buffered mode

pub fn lsee(self) -> &'a mut W

Lock access to SmartEEPROM data from any mean

pub fn usee(self) -> &'a mut W

Unlock access to SmartEEPROM data

pub fn lseer(self) -> &'a mut W

Lock access to the SmartEEPROM Register Address Space (above 64KB)

pub fn useer(self) -> &'a mut W

Unlock access to the SmartEEPROM Register Address Space (above 64KB)

pub unsafe fn bits(self, value: u8) -> &'a mut W

Writes raw bits to the field