Type Definition atsamd51n::qspi::instrframe::WIDTH_R

type WIDTH_R = R<u8, WIDTH_A>;

Reader of field WIDTH

Implementations

impl WIDTH_R

pub fn variant(&self) -> Variant<u8, WIDTH_A>

Get enumerated values variant

pub fn is_single_bit_spi(&self) -> bool

Checks if the value of the field is SINGLE_BIT_SPI

pub fn is_dual_output(&self) -> bool

Checks if the value of the field is DUAL_OUTPUT

pub fn is_quad_output(&self) -> bool

Checks if the value of the field is QUAD_OUTPUT

pub fn is_dual_io(&self) -> bool

Checks if the value of the field is DUAL_IO

pub fn is_quad_io(&self) -> bool

Checks if the value of the field is QUAD_IO

pub fn is_dual_cmd(&self) -> bool

Checks if the value of the field is DUAL_CMD

pub fn is_quad_cmd(&self) -> bool

Checks if the value of the field is QUAD_CMD