Type Definition atsamd51n::sdhc0::bsr::BOUNDARY_R

type BOUNDARY_R = R<u8, BOUNDARY_A>;

Reader of field BOUNDARY

Implementations

impl BOUNDARY_R

pub fn variant(&self) -> BOUNDARY_A

Get enumerated values variant

pub fn is_4k(&self) -> bool

Checks if the value of the field is _4K

pub fn is_8k(&self) -> bool

Checks if the value of the field is _8K

pub fn is_16k(&self) -> bool

Checks if the value of the field is _16K

pub fn is_32k(&self) -> bool

Checks if the value of the field is _32K

pub fn is_64k(&self) -> bool

Checks if the value of the field is _64K

pub fn is_128k(&self) -> bool

Checks if the value of the field is _128K

pub fn is_256k(&self) -> bool

Checks if the value of the field is _256K

pub fn is_512k(&self) -> bool

Checks if the value of the field is _512K