Type Definition atsamd51n::sdhc0::hc2r::SLCKSEL_R

type SLCKSEL_R = R<bool, SLCKSEL_A>;

Reader of field SLCKSEL

Implementations

impl SLCKSEL_R

pub fn variant(&self) -> SLCKSEL_A

Get enumerated values variant

pub fn is_fixed(&self) -> bool

Checks if the value of the field is FIXED

pub fn is_tuned(&self) -> bool

Checks if the value of the field is TUNED