Type Definition atsamd51n::sercom0::usart_int::ctrlb::CHSIZE_R

type CHSIZE_R = R<u8, CHSIZE_A>;

Reader of field CHSIZE

Implementations

impl CHSIZE_R

pub fn variant(&self) -> Variant<u8, CHSIZE_A>

Get enumerated values variant

pub fn is_8_bit(&self) -> bool

Checks if the value of the field is _8_BIT

pub fn is_9_bit(&self) -> bool

Checks if the value of the field is _9_BIT

pub fn is_5_bit(&self) -> bool

Checks if the value of the field is _5_BIT

pub fn is_6_bit(&self) -> bool

Checks if the value of the field is _6_BIT

pub fn is_7_bit(&self) -> bool

Checks if the value of the field is _7_BIT