Module atsamd51n::supc::bod33

BOD33 Control

Structs

ACTION_R

Field ACTION reader - Action when Threshold Crossed

ACTION_W

Field ACTION writer - Action when Threshold Crossed

BOD33_SPEC

BOD33 Control

ENABLE_R

Field ENABLE reader - Enable

ENABLE_W

Field ENABLE writer - Enable

HYST_R

Field HYST reader - Hysteresis value

HYST_W

Field HYST writer - Hysteresis value

LEVEL_R

Field LEVEL reader - Threshold Level for VDD

LEVEL_W

Field LEVEL writer - Threshold Level for VDD

PSEL_R

Field PSEL reader - Prescaler Select

PSEL_W

Field PSEL writer - Prescaler Select

R

Register BOD33 reader

RUNBKUP_R

Field RUNBKUP reader - Run in Backup mode

RUNBKUP_W

Field RUNBKUP writer - Run in Backup mode

RUNHIB_R

Field RUNHIB reader - Run in Hibernate mode

RUNHIB_W

Field RUNHIB writer - Run in Hibernate mode

RUNSTDBY_R

Field RUNSTDBY reader - Run in Standby mode

RUNSTDBY_W

Field RUNSTDBY writer - Run in Standby mode

STDBYCFG_R

Field STDBYCFG reader - Configuration in Standby mode

STDBYCFG_W

Field STDBYCFG writer - Configuration in Standby mode

VBATLEVEL_R

Field VBATLEVEL reader - Threshold Level in battery backup sleep mode for VBAT

VBATLEVEL_W

Field VBATLEVEL writer - Threshold Level in battery backup sleep mode for VBAT

W

Register BOD33 writer

Enums

ACTION_A

Action when Threshold Crossed

PSEL_A

Prescaler Select