Type Definition atsamd51n::wdt::config::PER_R

type PER_R = R<u8, PER_A>;

Reader of field PER

Implementations

impl PER_R

pub fn variant(&self) -> Variant<u8, PER_A>

Get enumerated values variant

pub fn is_cyc8(&self) -> bool

Checks if the value of the field is CYC8

pub fn is_cyc16(&self) -> bool

Checks if the value of the field is CYC16

pub fn is_cyc32(&self) -> bool

Checks if the value of the field is CYC32

pub fn is_cyc64(&self) -> bool

Checks if the value of the field is CYC64

pub fn is_cyc128(&self) -> bool

Checks if the value of the field is CYC128

pub fn is_cyc256(&self) -> bool

Checks if the value of the field is CYC256

pub fn is_cyc512(&self) -> bool

Checks if the value of the field is CYC512

pub fn is_cyc1024(&self) -> bool

Checks if the value of the field is CYC1024

pub fn is_cyc2048(&self) -> bool

Checks if the value of the field is CYC2048

pub fn is_cyc4096(&self) -> bool

Checks if the value of the field is CYC4096

pub fn is_cyc8192(&self) -> bool

Checks if the value of the field is CYC8192

pub fn is_cyc16384(&self) -> bool

Checks if the value of the field is CYC16384